blk-iocost: make ioc_now->now and ioc->period_at 64bit
commit ce95570acf741ea306baddcb43aba0b59b920a21 upstream.

Note: there are a few conflics need to resolved when the origin patch
applied.

They are in microseconds and wrap in around 1.2 hours with u32. While
unlikely, confusions from wraparounds are still possible. We aren't saving
anything meaningful by keeping these u32. Let's make them u64.

+++ block/blk-iocost.c
@@ -412,7 +412,7 @@ struct ioc {
 	atomic64_t			vtime_rate;
 
 	seqcount_t			period_seqcount;
-	u32				period_at;	/* wallclock starttime */
+	u64				period_at;	/* wallclock starttime */
 	u64				period_at_vtime; /* vtime starttime */
 
 	atomic64_t			cur_period;	/* inc'd each period */
@@ -511,7 +511,7 @@ struct ioc_cgrp {
 
 struct ioc_now {
 	u64				now_ns;
-	u32				now;
+	u64				now;
 	u64				vnow;
 	u64				vrate;
 };
